Haraka 2.8 (#46)
* updates for Haraka@2.8.0

* add sed backup empty extension
* haraka version alpha -> 2.8.3
* minecraft updates
* turn off attempts to install shellcheck
* base: only create ssl dirs if non-existent
* base: update installed packages
* spamassassin: add min/max spare settings to rc.conf
* vpopmail: prefix ch* cmds with stage_exec
* webmail: remove cracklib from qmailadmin (broken)

+++ provision-haraka.sh
@@ -7,16 +7,10 @@ export JAIL_START_EXTRA="devfs_ruleset=7"
 export JAIL_CONF_EXTRA="
 		devfs_ruleset = 7;"
 
+PREVIOUS_CONF="$ZFS_JAIL_MNT/haraka/usr/local/haraka/config"
 HARAKA_CONF="$STAGE_MNT/usr/local/haraka/config"
 
-install_haraka()
-{
-	tell_status "installing node & npm"
-	stage_pkg_install node npm gmake || exit
-
-	tell_status "installing Haraka"
-	stage_exec npm install -g Haraka ws express || exit
-
+haraka_github_updates() {
 	tell_status "updating files from GitHub repo"
 	local _ghi="$STAGE_MNT/usr/local/lib/node_modules/Haraka"
 	local _ghu='https://raw.githubusercontent.com/haraka/Haraka/master'
@@ -29,6 +23,18 @@ install_haraka()
 	done
 }
 
+install_haraka()
+{
+	tell_status "installing node & npm"
+	stage_pkg_install node npm gmake || exit
+	#stage_exec make -C /usr/ports/www/npm install clean
+
+	tell_status "installing Haraka"
+	stage_exec npm install -g Haraka ws express || exit
+
+	#haraka_github_updates
+}
+
 install_geoip_dbs()
 {
 	if ! zfs_filesystem_exists "$ZFS_DATA_VOL/geoip"; then
@@ -37,10 +43,9 @@ install_geoip_dbs()
 	fi
 
 	tell_status "enabling Haraka geoip plugin"
-	mkdir -p "$STAGE_MNT/usr/local/share/GeoIP"
-	sed -i -e 's/^;calc_distance=false/calc_distance=true/' "$HARAKA_CONF/connect.geoip.ini"
-	sed -i -e 's/^# connect.geoip/connect.geoip/' "$HARAKA_CONF/plugins"
+	sed -i .bak -e 's/^# connect.geoip/connect.geoip/' "$HARAKA_CONF/plugins"
 
+	mkdir -p "$STAGE_MNT/usr/local/share/GeoIP"
 	JAIL_CONF_EXTRA="$JAIL_CONF_EXTRA
 		mount += \"$ZFS_DATA_MNT/geoip \$path/usr/local/share/GeoIP nullfs ro 0 0\";"
 }
@@ -77,7 +82,7 @@ install_p0f()
 
 	get_public_facing_nic
 	if [ "$PUBLIC_NIC" != "bce1" ]; then
-		sed -i -e "s/ bce1 / $PUBLIC_NIC /" "$_start" || exit
+		sed -i '' -e "s/ bce1 / $PUBLIC_NIC /" "$_start" || exit
 	fi
 
 	stage_sysrc p0f_enable=YES
@@ -87,24 +92,41 @@ install_p0f()
 config_haraka_syslog()
 {
 	tell_status "switch Haraka logging to syslog"
-	sed -i -e 's/# log.syslog$/log.syslog/' "$HARAKA_CONF/plugins"
-	# sed -i -e 's/^daemon_log_file=.*/daemon_log_file=\/dev\/null/' "$HARAKA_CONF/smtp.ini"
-	sed -i -e 's/always_ok=false/always_ok=true/' "$HARAKA_CONF/log.syslog.ini"
+	sed -i '' -e 's/# log.syslog$/log.syslog/' "$HARAKA_CONF/plugins"
+
+	# send haraka logs to /dev/null
+	sed -i '' -e 's/^daemon_log_file=.*/daemon_log_file=\/dev\/null/' "$HARAKA_CONF/smtp.ini"
+
+	# don't write to daemon_log_file if syslog write was successful
+	echo "[general]
+always_ok=true" | tee -a  "$HARAKA_CONF/log.syslog.ini"
+}
+
+config_haraka_smtp_forward()
+{
+	tell_status "configure smtp forward to vpopmail jail"
+	if [ -f "$PREVIOUS_CONF/smtp_forward.ini" ]; then
+		tell_status "preserving existing smtp_forward.ini"
+		cp "$PREVIOUS_CONF/smtp_forward.ini" "$HARAKA_CONF/smtp_forward.ini"
+	else
+		echo "host=$(get_jail_ip vpopmail)
+port=25
+" | tee -a "$HARAKA_CONF/smtp_forward.ini"
+	fi
 }
 
 config_haraka_vpopmail()
 {
-	tell_status "configure smtp forward to vpopmail jail"
-	sed -i .bak \
-		-e "s/^host=localhost/host=$(get_jail_ip vpopmail)/" \
-		-e 's/^port=2555/port=25/' \
-		"$HARAKA_CONF/smtp_forward.ini"
-
 	tell_status "config SMTP AUTH using vpopmaild"
-	echo "host=$(get_jail_ip vpopmail)" > "$HARAKA_CONF/auth_vpopmaild.ini"
+	if [ -f "$PREVIOUS_CONF/auth_vpopmaild.ini" ]; then
+		tell_status "preserving existing auth_vpopmaild.ini"
+		cp "$PREVIOUS_CONF/auth_vpopmaild.ini" "$HARAKA_CONF/auth_vpopmaild.ini"
+	else
+		echo "host=$(get_jail_ip vpopmail)" > "$HARAKA_CONF/auth_vpopmaild.ini"
+	fi
 
 	# shellcheck disable=1004
-	sed -i -e '/^# auth\/auth_ldap$/a\
+	sed -i '' -e '/^# auth\/auth_ldap$/a\
 auth\/auth_vpopmaild
 ' "$HARAKA_CONF/plugins"
 }
@@ -112,7 +134,16 @@ auth\/auth_vpopmaild
 config_haraka_qmail_deliverable()
 {
 	tell_status "config recipient validation with Qmail::Deliverable"
-	sed -i -e "s/^host=127.0.0.1/host=$(get_jail_ip vpopmail)/" "$HARAKA_CONF/rcpt_to.qmail_deliverable.ini"
+	if [ -f "$PREVIOUS_CONF/rcpt_to.qmail_deliverable.ini" ]; then
+		tell_status "preserving existing rcpt_to.qmail_deliverable.ini"
+		cp "$PREVIOUS_CONF/rcpt_to.qmail_deliverable.ini" \
+		    "$HARAKA_CONF/rcpt_to.qmail_deliverable.ini"
+	else
+		echo "check_outbound=true
+host=$(get_jail_ip vpopmail)" | \
+		tee -a "$HARAKA_CONF/rcpt_to.qmail_deliverable.ini"
+	fi
+
 	sed -i .bak \
 		-e 's/^#rcpt_to.qmail_deliverable/rcpt_to.qmail_deliverable/' \
 		-e 's/^rcpt_to.in_host_list/# rcpt_to.in_host_list/' \
@@ -124,7 +155,7 @@ config_haraka_p0f()
 	install_p0f
 
 	tell_status "enable Haraka p0f plugin"
-	sed -i -e 's/^# connect.p0f/connect.p0f/' "$HARAKA_CONF/plugins"
+	sed -i '' -e 's/^# connect.p0f/connect.p0f/' "$HARAKA_CONF/plugins"
 }
 
 config_haraka_spamassassin()
@@ -134,15 +165,21 @@ config_haraka_spamassassin()
 		return
 	fi
 
-	tell_status "configuring Haraka spamassassin plugin"
-	sed -i.bak \
-		-e "/^spamd_socket/ s/127.0.0.1:/$(get_jail_ip spamassassin):/" \
-		-e 's/^;spamd_user=$/spamd_user=first-recipient/' \
-		-e '/^; reject_threshold/ s/^; //' \
-		-e 's/^; relay_reject_threshold$/relay_reject_threshold/' \
-		"$HARAKA_CONF/spamassassin.ini"
+	tell_status "enabling Haraka spamassassin plugin"
+	sed -i '' -e 's/^#spamassassin$/spamassassin/' "$HARAKA_CONF/plugins"
 
-	sed -i -e 's/^#spamassassin$/spamassassin/' "$HARAKA_CONF/plugins"
+	if [ -f "$PREVIOUS_CONF/spamassassin.ini" ]; then
+		tell_status "preserving existing spamassassin.ini"
+		cp "$PREVIOUS_CONF/spamassassin.ini" "$HARAKA_CONF/spamassassin.ini"
+	else
+		tell_status "configuring Haraka spamassassin plugin"
+		echo "spamd_socket=$(get_jail_ip spamassassin):783
+old_headers_action=rename
+spamd_user=first-recipient
+reject_threshold=10
+relay_reject_threshold=7
+" | tee -a "$HARAKA_CONF/spamassassin.ini"
+	fi
 }
 
 config_haraka_avg()
@@ -158,13 +195,12 @@ config_haraka_avg()
 	JAIL_CONF_EXTRA="$JAIL_CONF_EXTRA
 		mount += \"$ZFS_DATA_MNT/avg \$path/data/avg nullfs rw 0 0\";"
 
-	sed -i .bak \
-		-e "s/;host.*/host = $(get_jail_ip avg)/" \
-		-e 's/;tmpdir.*/tmpdir=\/data\/avg/' \
-		"$HARAKA_CONF/avg.ini"
+	echo "host = $(get_jail_ip avg)
+tmpdir=/data/avg
+" | tee -a "$HARAKA_CONF/avg.ini"
 
 	# shellcheck disable=1004
-	sed -i -e '/clamd$/a\
+	sed -i '' -e '/clamd$/a\
 avg
 ' "$HARAKA_CONF/plugins"
 }
@@ -176,14 +212,29 @@ config_haraka_clamav()
 		return
 	fi
 
+	tell_status "enabling Haraka clamav plugin"
+	sed -i '' -e 's/^#clamd$/clamd/' "$HARAKA_CONF/plugins"
+
 	tell_status "configure Haraka clamav plugin"
-	echo "clamd_socket=$(get_jail_ip clamav):3310" >> "$HARAKA_CONF/clamd.ini"
-	sed -i -e 's/^#clamd$/clamd/' "$HARAKA_CONF/plugins"
+	echo "clamd_socket=$(get_jail_ip clamav):3310
+
+[reject]
+virus=true
+error=false
+DetectBrokenExecutables=false
+Structured=false
+ArchiveBlockEncrypted=false
+PUA=false
+OLE2=false
+Safebrowsing=false
+UNOFFICIAL=false
+Phishing=false
+" | tee -a "$HARAKA_CONF/clamd.ini"
 }
 
-config_haraka_tls() {	
+config_haraka_tls() {
 	tell_status "enable TLS encryption"
-	sed -i -e 's/^# tls$/tls/' "$HARAKA_CONF/plugins"
+	sed -i '' -e 's/^# tls$/tls/' "$HARAKA_CONF/plugins"
 	cp /etc/ssl/certs/server.crt "$HARAKA_CONF/tls_cert.pem"
 	cp /etc/ssl/private/server.key "$HARAKA_CONF/tls_key.pem"
 }
@@ -191,31 +242,11 @@ config_haraka_tls() {
 config_haraka_dnsbl()
 {
 	tell_status "configuring dnsbls"
-	echo 'reject=0' > "$HARAKA_CONF/dnsbl.ini"
-	echo 'periodic_checks=30' >> "$HARAKA_CONF/dnsbl.ini"
-
-	echo "; zen.spamhaus.org
-b.barracudacentral.org
-truncate.gbudb.net
-cbl.abuseat.org
-psbl.surriel.com
-bl.spamcop.net
-dnsbl-1.uceprotect.net
-pbl.spamhaus.org
-xbl.spamhaus.org
-" | tee "$HARAKA_CONF/dnsbl.zones"
-}
-
-cleanup_deprecated_haraka()
-{
-	rm "$HARAKA_CONF/lookup_rdns.strict.ini"
-	rm "$HARAKA_CONF/lookup_rdns.strict.timeout"
-	rm "$HARAKA_CONF/lookup_rdns.strict.whitelist"
-	rm "$HARAKA_CONF/lookup_rdns.strict.whitelist_regex"
-	rm "$HARAKA_CONF/mail_from.access.blacklist_regex"
-	rm "$HARAKA_CONF/rcpt_to.blocklist"
-	rm "$HARAKA_CONF/rdns.allow_regexps"
-	rm "$HARAKA_CONF/rdns.deny_regexps"
+	echo 'reject=false
+search=all
+enable_stats=false
+zones=b.barracudacentral.org, truncate.gbudb.net, psbl.surriel.com, bl.spamcop.net, dnsbl-1.uceprotect.net, zen.spamhaus.org, dnsbl.sorbs.net, dnsbl.justspam.org, bad.psky.me
+' | tee -a "$HARAKA_CONF/dnsbl.ini"
 }
 
 config_haraka_rspamd()
@@ -226,13 +257,12 @@ config_haraka_rspamd()
 	fi
 
 	tell_status "configure Haraka rspamd plugin"
-	sed -i .bak \
-		-e "s/;host.*/host = $(get_jail_ip rspamd)/" \
-		-e 's/;always_add_headers = false/always_add_headers = true/' \
-		"$HARAKA_CONF/rspamd.ini" || exit
+	echo "host = $(get_jail_ip rspamd)
+always_add_headers = true
+" | tee -a "$HARAKA_CONF/rspamd.ini" || exit
 
 	# shellcheck disable=1004
-	sed -i -e '/spamassassin$/a\
+	sed -i '' -e '/spamassassin$/a\
 rspamd
 ' "$HARAKA_CONF/plugins" || exit
 }
@@ -274,6 +304,14 @@ config_haraka_plugins()
 		"$HARAKA_CONF/plugins"
 }
 
+config_install_default()
+{
+	local _source="$STAGE_MNT/usr/local/lib/node_modules/Haraka/config"
+	local _dest="$STAGE_MNT/usr/local/haraka/config"
+	echo "cp $_source/$1 $_dest/$1"
+	cp "$_source/$1" "$_dest/$1"
+}
+
 config_haraka_limit()
 {
 	tell_status "configuring limit plugin"
@@ -281,6 +319,7 @@ config_haraka_limit()
 		-e 's/^max_unrecognized_commands/limit/' \
 		"$HARAKA_CONF/plugins"
 
+	config_install_default limit.ini
 	sed -i .bak \
 		-e 's/^; max/max/' \
 		-e 's/^; history/history/' \
@@ -291,14 +330,14 @@ config_haraka_limit()
 
 config_haraka_dkim()
 {
-	sed -i .bak -e 's/^disabled = true/disabled = false/' \
-		"$HARAKA_CONF/dkim_sign.ini"
+	echo 'disabled=false' | tee -a "$HARAKA_CONF/dkim_sign.ini"
 
-	_dkim_dir="$ZFS_JAIL_MNT/haraka/usr/local/haraka/config/dkim/$TOASTER_MAIL_DOMAIN"
-	if [ -d "$_dkim_dir" ]; then
+	mkdir -p "$HARAKA_CONF/dkim"
+	config_install_default "dkim/dkim_key_gen.sh"
+
+	if [ -d "$PREVIOUS_CONF/dkim" ]; then
 		tell_status "copying active DKIM keys"
-		cp -R "$_dkim_dir/" "$HARAKA_CONF/dkim/"
-		return
+		cp -R "$PREVIOUS_CONF/dkim/" "$HARAKA_CONF/dkim/"
 	else
 		tell_status "generating DKIM keys"
 		cd "$HARAKA_CONF/dkim" || exit
@@ -312,11 +351,22 @@ config_haraka_dkim()
 
 config_haraka_karma()
 {
-	sed -i .bak \
-		-e '/^dbid/ s/= 0/= 1/' \
-		-e "/^server_ip/ s/127.0.0.1/$(get_jail_ip redis)/" \
-		-e '/^plugins=/ s/clamd, //' \
-		"$HARAKA_CONF/karma.ini"
+	if [ -d "$PREVIOUS_CONF/karma.ini" ]; then
+		tell_status "preserving karma.ini"
+		cp "$PREVIOUS_CONF/karma.ini" "$HARAKA_CONF/karma.ini"
+		return
+	fi
+
+	tell_status "configuring karma plugin"
+	echo "
+[redis]
+dbid=1
+server_ip=$(get_jail_ip redis)
+
+[deny_excludes]
+plugins=send_email, access, helo.checks, data.headers, mail_from.is_resolvable, avg, limit, attachment, tls
+" | tee -a "$HARAKA_CONF/karma.ini"
+
 }
 
 config_haraka_redis()
@@ -332,16 +382,16 @@ EO_REDIS_CONF
 }
 
 config_haraka_geoip() {
-	sed -i .bak \
-		-e '/^;[asn]/ s/^;//' \
-		-e '/^;report_as=/ s/^;//' \
-		"$HARAKA_CONF/connect.geoip.ini"
+	echo "calc_distance=true
+[asn]
+report_as=connect.asn
+" | tee -a "$HARAKA_CONF/connect.geoip.ini"
 }
 
 config_haraka_http()
 {
 	tell_status "enable Haraka HTTP server"
-	sed -i -e 's/; listen=\[::\]:80/listen=0.0.0.0:80/' "$HARAKA_CONF/http.ini"
+	echo "listen=0.0.0.0:80" | tee -a "$HARAKA_CONF/http.ini"
 }
 
 configure_haraka()
@@ -353,15 +403,17 @@ configure_haraka()
 	echo 'LOGINFO' > "$HARAKA_CONF/loglevel"
 	echo '3' > "$HARAKA_CONF/tarpit.timeout"
 
+	config_install_default plugins
 	config_haraka_smtp_ini
 	config_haraka_plugins
 	config_haraka_limit
 	config_haraka_syslog
 	config_haraka_vpopmail
+	config_haraka_smtp_forward
 	config_haraka_qmail_deliverable
 	config_haraka_dnsbl
 
-	sed -i -e 's/^; reject=.*/reject=no/' "$HARAKA_CONF/data.headers.ini"
+	echo "reject=no" | tee -a "$HARAKA_CONF/data.headers.ini"
 
 	config_haraka_http
 	config_haraka_tls
@@ -377,7 +429,6 @@ configure_haraka()
 	config_haraka_geoip
 
 	install_geoip_dbs
-	cleanup_deprecated_haraka
 }
 
 start_haraka()
